Update README.md
Designed and built the College Scheduler App, proposed to help college students by providing a straightforward and efficient way to manage their academic schedules. This app focuses on core functionalities that allow students to track their classes, assignments, and exams. The aim was to offer a user-friendly and accessible tool that adapts to various class formats and individual schedules, helping students stay organized and focused on their academic commitments.